+++ CHANGES.md
@@ -3,6 +3,9 @@ E-Maj - Change log
 <devel>
 ------
 ###Enhancements:###
+  * Record the E-Maj rollback start, initialization end and locking end
+    timestamps into the emaj_rlbk table and let the emaj_rollback_activity()
+    function report both rollback planning and locking phases duration.
   * Minor code changes.
 
 ###Bug fixes:###